backend-django/requirements.txt
django==4.2.20
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
django-cors-headers==4.3.1
django-filter==23.5
bcrypt==4.1.2
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
openpyxl==3.1.2
Pillow==10.2.0
